requirements.txt
Flask
Flask-WTF
Flask-Mail
passlib
Werkzeug
flask-paginate
spacy
python-dotenv
pymongo
mysql-connector-python